Have BaseCsc report its initial summary state.
Add argument initial_state to BaseCsc, in order to support TestCsc
and testing in general. This is part of reporting initial state,
because otherwise `BaseCsc` will report STANDBY even if the subclass
(such as TestCsc) sets a different initial state.
Also allow state values to be integers and add tests for this.