Simplify logging
Replace the queue log handler with a new handler that simply writes
the message directly. This makes it easier to shut down without losing
the last log message (though one still has to wait a bit for SAL
to write final topics, as always), eliminates an annoying delay
in output of log messages, and greatly simplifies the code.