Fix test_object issues
There were instances where RemoteCommand.start was not awaited;
that worked, but is not something users should rely on and i am about to change it to not work.
Also a timing test was too picky; I added a bit of slop.
Finally, test read data before checking that the read queue is empty,
in case the check fails and we want to examine the data read by the
function that should have timed out. This can occur when running the tests
with pytest -n 4, and I don't know why. The returned data appears to be
the last value sent.