Add HOME and OSPL_HOME to tests/SConscript
and remove OPENSPLICE_LOC

Environment variable HOME is wanted by astropy and
test_script_schema_process in test_base_script.py was failing
because a warning was showing up in stderr when it was not found.
Environment variable OSPL_HOME is sometimes wanted by OpenSplice;
I found an error message about it in ospl-error.log.
Environment variable OPENSPLICE_LOC is not needed.