Clarify BaseCsc.main(index=0, ...) behavior
which matches `None`, as usual for salobj.
Modify the documentation and code to make it clearer
and update the unit test.
Also stop mentioning `False` as a value; it will continue
to work as `None` or `0`, but it doesn't match any other
usage and was not used anywhere (according to a global search).

Update the unit test to test `0` and stop testing `False`.