Improve error handling in BaseCsc.fault
Log the problem and continue if the error code is not an integer.
Log the exception and continue if `BaseCsc.report_summary_state` raises.
Avoid an infinite loop if `report_summary_state` calls `fault`
(which it might do if something goes wrong).
Deprecate `code=None`.
Update `TestCsc` to specify a code when calling `fault`.
Add tests for these changes.