Improve test_speed.py
Move creation of the remote and topic writer subprocess to a context manager,
to simplify the code and make sure we close the topic writer when done.
Fix measurement of topic creation speed (the inverse was reported).
Improve output.
Do not fail if the read speed measurement loses any samples;
such loss is likely because writing is faster than reading.
Report the number of samples lost in the read test.