Fix two tests in test_topics.py to pass if LSST_DDS_IP is defined
test_command_isolation as assuming that LSST_DDS_IP was not defined,
so that it could create two Domains the host would be different.
Fix the test to not make this assumption.

Also test_remote_command_set was needlessly creating two domains
and then testing that they had different host values.