Improve command callback tests a bit
Move the csc creation out of check_controller_command_callback_failure
so that the test of returning a failed ackcmd can use it as well.
Rename the test methods for clarity.
Expand the test of setting and getting controller command callbacks
to include clearing the callback, and expand the comments.