Add type annotations
to the library code and to some (but not all) unit tests.
Tests that work with CSCs, Controllers and Remotes aren't worth annotating
because mypy doesn't know about the dynamically added SAL/DDS topics.
The resulting code would have so many ``# type: ignore`` lines
that it would be hard to read.