SalLogHandler: reduce 'SalLogHandler._async_emit' warnings
at shutdown as follows:
* Create a task when run from the foreground thread
* Create a concurrent Future when run from a background thread
* Keep track of unfinished tasks and futures and cancel them
  when the handler is closed.

Note that, for reasons I do not understand, messages logged
from background threads still cause these warnings. Cancelling
the unfinished futures does not help (and in fact they are already
done when close is called). So the only change that actually helps
is to use tasks when run from a foreground thread (and cancelling
them in close, but since Controller pauses at shutdown to give SAL
messages time to be sent, so there are never any unfinished tasks).