In ``sal_info.py``, add new environment variable to allow specifying the replication factor when creating topics.
By default the replication factor is 1, which is suitable for testing purposes.
For production and test stand deployments this can be set to an appropriate value, e.g. 3.
Since topics are only created once, if a topic was created with a replication factor, and one wants to update it, the topic needs to be deleted for any change to take effect.