Add WildcardIndexError to handle wildcard indices in SAL
Add a custom WildcardIndexError exception to handle cases where the
index in a SAL component name is a wildcard ('*'), such as  "Test:*".

This is useful for situations where having a wildcard index is valid,
for example, when applying a state or action to all indices of a given
CSC name.

See application of this in ts_standardscripts set_summary_state.py.